+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+[1.9.1] (2025-07-01)
+--------------------
+
 ### Fixed
 
 - Fixed Whisperx diarization pipeline initialization
@@ -303,6 +306,7 @@ Unreleased
 - mp3 support by using FFmpeg instead of librosa in #8
 - add language detection endpoint in #9
 
+[1.9.1]: https://github.com/ahmetoner/whisper-asr-webservice/releases/tag/v1.9.1
 [1.9.0]: https://github.com/ahmetoner/whisper-asr-webservice/releases/tag/v1.9.0
 [1.8.2]: https://github.com/ahmetoner/whisper-asr-webservice/releases/tag/v1.8.2
 [1.8.1]: https://github.com/ahmetoner/whisper-asr-webservice/releases/tag/v1.8.1